# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Permanent Portfolio Planner API
  version: 1.0.0
servers:
  - url: http://localhost:3001
paths:
  /health:
    get:
      summary: Service health
      responses:
        "200":
          description: Healthy
  /api/auth/register:
    post:
      summary: Register an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "201":
          description: Registered
        "400":
          description: Validation error
        "409":
          description: Email already exists
  /api/auth/verify-email:
    post:
      summary: Verify a newly registered email address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
                  pattern: '^\d{6}$'
      responses:
        "200":
          description: Email verified
        "400":
          description: Invalid or expired verification code
  /api/auth/resend-verification:
    post:
      summary: Send a new email verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Request accepted
  /api/auth/login:
    post:
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Access token
        "401":
          description: Invalid credentials
        "403":
          description: Email has not been verified
  /api/auth/change-password:
    post:
      summary: Change the signed-in user's password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 10
      responses:
        "200":
          description: Password changed and a replacement access token issued
        "401":
          description: Current password is incorrect
  /api/data:
    get:
      summary: Read the current user's portfolio backup
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Portfolio payload
        "401":
          description: Unauthorized
    put:
      summary: Replace the current user's portfolio backup
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payload]
              properties:
                payload:
                  type: object
      responses:
        "200":
          description: Saved
        "400":
          description: Invalid portfolio payload
        "401":
          description: Unauthorized
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 10